Refactor ctypeswriter, nuke __init_methods__
Using methods defined at class level now, with lazily evaluated argument
lists - in other words, they are still passed as strings for the sake of
being able to reference to names that aren't available at the point
where the method is defined.

Each register line looks like this now:

    get_size, r_cons_get_size = \
        register('r_cons_get_size', 'POINTER(c_int)','c_int')

"register" returns both the api method and the r_ method. The former
wraps the latter, and the latter wraps the ctypes function slightly.

There's also a new AddressHolder class for the _o attribute. It holds
addresses. The idea is to not require to go through __init__ to have it.

Also blacklisted the "del" keyword in get_alias.